requirements.txt
langchain-community>=0.2.11
tavily-python
gitpython
selenium>=4.0.0
python-dotenv>=1.0.0
atproto>=0.0.31
langgraph
langgraph-sdk
langgraph-cli
langgraph-checkpoint-sqlite
langsmith
langchain-core
langchain-community
langchain-openai
notebook
pydantic
langchain_anthropic
wikipedia